Reconfigure modified targets after successful runs
This allows tasks to be created that compose newer-prefixed tasks with other tasks that might be configured with Grunt's template syntax to use the same config.  For example a task called `lint-then-watch` might be registered that composes `newer:jshint` and `watch`.  If the `watch` targets use template syntax to get the file config from the `jshint` targets, the `jshint` targets must be reconfigured after being run with a modified configuration.