Pass around override function
The override function will call a user provided function for each file that the newer task determines is not newer.  The user provided function will have an opportunity to override this - essentially forcing a file to be considered newer even if its mtime doesn't show it.  This can be used in cases like LESS compilation where a file might have imports that have been updated.